Add support for loading syntax colours for other languages
Support loading colours for other languages
Introduces "inheritance" for colours so unspecified colours will be
set to match the colour of something else which was specified.
This means if a .prf is loaded which only includes MATLAB syntax colours, the
colours for other languages are automatically set appropriately.
Also, if the value for one particular text colour is ommitted, it will not be
left the same, potentially blending in with the background.