Add support for export/importing strings in additional languages
In particular, this is for the benefit of being able to set the Java method
font style.

The string preferences can be restricted to only be imported or exported if it
matches a target regex. This restriction can be lifted by setting the target to
be '' to match anything (including empty string) or '.' to match anything
except an empty string.

Closes #7.